[Bugfix] Kill orphaned EngineCores
If the main process exits and EngineCores are left orphaned, an observer process will kill the orphaned EngineCores and free memory. There are other alternative methods to be considered as well, such as running observer loops within the EngineCore subprocesses, but for now this fixes the issue. The observation interval is hardcoded as 20 seconds in CoreEngineProcManager._alive_check_interval.

Signed-off-by: Syed Muhammad Bin Asif <syedmba7@connect.hku.hk>