[Bugfix] Fix V1 AsyncLLM errored property and kill orphaned EngineCores
If the EngineCore subprocess dies, calling AsyncLLM.check_health() will raise an EngineDeadError.

If the main process exits and EngineCores are left orphaned, an observer process will kill the orphaned EngineCores and free memory. There are other alternative methods to be considered as well, such as running observer loops within the EngineCore subprocesses, but for now this fixes the issue. The observation interval is added as a new ParallelConfig variable called engine_core_orphaned_check_interval and CoreEngineProcManager.alive_check_interval takes on this value. The default interval is set to be 20 seconds.

We add relevant tests for these scenarios as well.

Signed-off-by: Syed Muhammad Bin Asif <syedmba7@connect.hku.hk>